pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "birdview"
version = "0.1.0"
description = "Minimal terminal Twitter/X client"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "tweepy>=4.14",
    "rich>=13.0",
    "click>=8.0",
    "python-dotenv>=1.0",
    "pyperclip>=1.8",
]

[project.scripts]
birdview = "birdview.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[tool.setuptools.packages.find]
include = ["birdview*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
